+++ src/specific/s_music.h
@@ -2,10 +2,13 @@
 
 #include "global/types.h"
 
-bool __cdecl S_Music_Init(void);
-void __cdecl S_Music_Shutdown(void);
-void __cdecl S_Music_Play(int16_t track_id, bool is_looped);
-void __cdecl S_Music_Stop(void);
-bool __cdecl S_Music_PlaySynced(int32_t track_id);
-uint32_t __cdecl S_Music_GetFrames(void);
-void __cdecl S_Music_SetVolume(int32_t volume);
+typedef struct S_MUSIC_BACKEND {
+    const char *name;
+    bool (*Init)(void);
+    void (*Shutdown)(void);
+    void (*Play)(int32_t track_id, bool is_looped);
+    void (*Stop)(void);
+    bool (*PlaySynced)(int32_t track_id);
+    uint32_t (*GetFrames)(void);
+    void (*SetVolume)(int32_t volume);
+} S_MUSIC_BACKEND;