+++ src/specific/s_music_mm.c
@@ -1,10 +1,19 @@
-#include "specific/s_music.h"
+#include "specific/s_music_mm.h"
 
 #include "global/funcs.h"
 #include "global/vars.h"
 #include "lib/winmm.h"
+#include "log.h"
 
-bool __cdecl S_Music_Init(void)
+static bool __cdecl S_Music_MM_Init(void);
+static void __cdecl S_Music_MM_Shutdown(void);
+static void __cdecl S_Music_MM_Play(int32_t track_id, bool is_looped);
+static void __cdecl S_Music_MM_Stop(void);
+static bool __cdecl S_Music_MM_PlaySynced(int32_t track_id);
+static uint32_t __cdecl S_Music_MM_GetFrames(void);
+static void __cdecl S_Music_MM_SetVolume(int32_t volume);
+
+static bool __cdecl S_Music_MM_Init(void)
 {
     MCI_OPEN_PARMS open_params;
     open_params.lpstrDeviceType = "cdaudio";
@@ -23,14 +32,14 @@ bool __cdecl S_Music_Init(void)
     return true;
 }
 
-void __cdecl S_Music_Shutdown(void)
+static void __cdecl S_Music_MM_Shutdown(void)
 {
     MCI_GENERIC_PARMS params;
     mciSendCommand(g_MciDeviceID, MCI_STOP, 0, (DWORD_PTR)&params);
     mciSendCommand(g_MciDeviceID, MCI_CLOSE, 0, (DWORD_PTR)&params);
 }
 
-void __cdecl S_Music_Play(int16_t track_id, bool is_looped)
+static void __cdecl S_Music_MM_Play(int32_t track_id, bool is_looped)
 {
     if (g_OptionMusicVolume == 0) {
         return;
@@ -52,7 +61,7 @@ void __cdecl S_Music_Play(int16_t track_id, bool is_looped)
     }
 }
 
-void __cdecl S_Music_Stop(void)
+static void __cdecl S_Music_MM_Stop(void)
 {
     if (g_CD_TrackID <= 0) {
         return;
@@ -64,7 +73,7 @@ void __cdecl S_Music_Stop(void)
     g_CD_LoopTrack = 0;
 }
 
-bool __cdecl S_Music_PlaySynced(int32_t track_id)
+static bool __cdecl S_Music_MM_PlaySynced(int32_t track_id)
 {
     g_CD_TrackID = track_id;
 
@@ -75,23 +84,20 @@ bool __cdecl S_Music_PlaySynced(int32_t track_id)
     if (mciSendCommand(
             g_MciDeviceID, MCI_SET, MCI_SET_TIME_FORMAT, (DWORD_PTR)&set_params)
         != 0) {
-        return false;
+        return FALSE;
     }
 
     MCI_PLAY_PARMS play_params;
     play_params.dwFrom = track_id;
     play_params.dwTo = track_id + 1;
-    if (mciSendCommand(
+    return (
+        mciSendCommand(
             g_MciDeviceID, MCI_PLAY, MCI_NOTIFY_FAILURE | MCI_NOTIFY_ABORTED,
             (DWORD_PTR)&play_params)
-        != 0) {
-        return false;
-    }
-
-    return true;
+        == 0);
 }
 
-uint32_t __cdecl S_Music_GetFrames(void)
+static uint32_t __cdecl S_Music_MM_GetFrames(void)
 {
     MCI_STATUS_PARMS status_params;
     status_params.dwItem = MCI_STATUS_POSITION;
@@ -109,7 +115,7 @@ uint32_t __cdecl S_Music_GetFrames(void)
     return (min * 60 + sec) * 75 + frame;
 }
 
-void __cdecl S_Music_SetVolume(int32_t volume)
+static void __cdecl S_Music_MM_SetVolume(int32_t volume)
 {
     int32_t aux_dev_count = auxGetNumDevs();
     if (aux_dev_count == 0) {
@@ -139,4 +145,20 @@ void __cdecl S_Music_SetVolume(int32_t volume)
     if (!is_volume_set && device_id != -1) {
         auxSetVolume(device_id, MAKELONG(volume, volume));
     }
+}
+
+static struct S_MUSIC_BACKEND m_Backend = {
+    .name = "MM",
+    .Init = S_Music_MM_Init,
+    .Shutdown = S_Music_MM_Shutdown,
+    .Play = S_Music_MM_Play,
+    .Stop = S_Music_MM_Stop,
+    .PlaySynced = S_Music_MM_PlaySynced,
+    .GetFrames = S_Music_MM_GetFrames,
+    .SetVolume = S_Music_MM_SetVolume,
+};
+
+struct S_MUSIC_BACKEND *S_Music_MM_Factory(void)
+{
+    return &m_Backend;
 }