+++ src/specific/s_music_pauld.h
@@ -1,0 +1,6 @@
+#pragma once
+
+#include "global/types.h"
+#include "specific/s_music.h"
+
+struct S_MUSIC_BACKEND *S_Music_PaulD_Factory(void);