+++ CHANGELOG.md
@@ -1,1 +1,2 @@
 ## [Unreleased](https://github.com/rr-/TR2X/compare/stable...develop) - ××××-××-××
+- fixed CDAudio not playing on certain versions (uses PaulD patch)